test(update): split update status coverage
